+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Version 1.4.2
 - fix VRAM leak when playing Bink video (RF bug)
 - fix AI flee animation regression
 - fix crash caused by AI state not being properly updated during level load for entites that are taken from the previous level
+- fix crash caused by corpse pose pointing to not loaded entity action animation for corpses that are taken from the previous level
 
 Version 1.4.1 (released 2019-12-07)
 -----------------------------------